Do not delete .gitkeep while cleaning uploads folder

+++ assetMG.py
@@ -518,14 +518,16 @@ def clean_dir():
   status=200
   folder = UPLOAD_FOLDER
   for filename in os.listdir(folder):
-      file_path = os.path.join(folder, filename)
-      try:
-          if os.path.isfile(file_path) or os.path.islink(file_path):
-              os.unlink(file_path)
-          elif os.path.isdir(file_path):
-              shutil.rmtree(file_path)
-      except Exception as e:
-          logging.error('Failed to delete %s. Reason: %s' % (file_path, e))
+    if filename == '.gitkeep':
+      continue
+    file_path = os.path.join(folder, filename)
+    try:
+      if os.path.isfile(file_path) or os.path.islink(file_path):
+        os.unlink(file_path)
+      elif os.path.isdir(file_path):
+        shutil.rmtree(file_path)
+    except Exception as e:
+      logging.error('Failed to delete %s. Reason: %s' % (file_path, e))
 
   if len(os.listdir(folder)) != 0:
     status=500
